chatbot/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backoff==2.2.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.0
chromadb-client==0.4.15.dev0
click==8.0.3
cloud-init==23.2.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.6.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.1.3
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gradio==3.47.1
gradio_client==0.6.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.20.2
httpx==0.25.0
huggingface-hub==0.18.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
importlib-resources==6.1.0
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.0.314
langsmith==0.0.43
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.8.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
openai==0.28.1
orjson==3.9.9
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pexpect==4.8.0
Pillow==10.0.1
posthog==3.0.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==1.10.13
pydub==0.25.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
python-multipart==0.0.6
pytz==2022.1
PyYAML==5.4.1
regex==2023.10.3
requests==2.31.0
SecretStorage==3.3.1
semantic-version==2.10.0
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sos==4.5.6
SQLAlchemy==2.0.22
ssh-import-id==5.11
starlette==0.27.0
systemd-python==234
tenacity==8.2.3
tiktoken==0.5.1
toolz==0.12.0
tqdm==4.66.1
Twisted==22.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.23.2
wadllib==1.3.6
websockets==11.0.3
yarl==1.9.2
zipp==1.0.0
zope.interface==5.4.0
